refactor(signin and signup): refactored sign in and sign up pages to match other page components
Flattened sign in and sign up pages and moved styles to root styles directory. Also moved Join Us
button into condition in Layout component
