+++ pages/users/signin.tsx
@@ -1,47 +1,24 @@
-import { useState } from 'react';
+import { signIn } from 'next-auth/client';
 import { Grid, TextField, Button } from '@material-ui/core';
 import { Formik, Form } from 'formik';
-import SignupSchema from 'interfaces/signup';
+import SigninSchema from 'interfaces/signin';
 import Layout from 'components/Layout';
-import { signIn } from 'next-auth/client';
-import styles from './signup.module.css';
+import styles from '../../styles/users/Signin.module.css';
 
 interface FormValues {
   email: string;
   password: string;
-  confirmPassword: string;
 }
 
 interface ErrorValues {
   email?: string;
   password?: string;
-  confirmPassword?: string;
 }
 
-// UserSignUp page. Will need additional email verification to be able to create organizations.
-const UserSignUp: React.FC = () => {
-  const [emailExists, setEmailExists] = useState(false);
-
+// UserSignIn page. Will need additional email verification to be able to create organizations.
+const UserSignIn: React.FC = () => {
   const handleSubmit = async (values: FormValues): Promise<any> => {
-    // Make post request
-    const res: any = await fetch('/api/users/', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        email: values.email,
-        password: values.password,
-      }),
-    });
-
-    // Check if email exists
-    if (!res) {
-      setEmailExists(true);
-      return new Error('Email already exists.');
-    }
-
-    // Sign in user
+    // Sign in with credentials
     try {
       signIn('credentials', {
         email: values.email,
@@ -51,13 +28,12 @@ const UserSignUp: React.FC = () => {
     } catch {
       throw new Error('Could not sign in.');
     }
-    return res;
   };
 
   // Validates inputs
   const validate = (values: FormValues): ErrorValues => {
     const errors: { [k: string]: string } = {};
-    const { error } = SignupSchema.validate(values, {
+    const { error } = SigninSchema.validate(values, {
       context: { strict: true },
       abortEarly: false,
     });
@@ -73,14 +49,13 @@ const UserSignUp: React.FC = () => {
   };
 
   const constructRow = (
-    title: string,
     varName: string,
     handleChange: any,
     error?: string
   ): any => (
     <>
       <Grid item xs={4}>
-        <div className={styles.entryName}>* {title}</div>
+        <div className={styles.entryName}>{varName}</div>
       </Grid>
       <Grid item xs={5}>
         <TextField
@@ -91,7 +66,7 @@ const UserSignUp: React.FC = () => {
           variant="outlined"
           onChange={handleChange}
           id={varName}
-          label={title}
+          label={varName}
           type={
             varName === 'password' || varName === 'confirmPassword'
               ? 'password'
@@ -104,11 +79,11 @@ const UserSignUp: React.FC = () => {
   );
 
   return (
-    <Layout title="Sign Up">
+    <Layout title="Sign In">
       <div className={styles.wrapper}>
         <div className={styles.titles}>
-          <h1>Join Us!</h1>
-          <h2>Add your organization!</h2>
+          <h1>Welcome Back!</h1>
+          <h2>Organization Log In</h2>
           &nbsp;
         </div>
         <div className={styles.entries}>
@@ -116,7 +91,6 @@ const UserSignUp: React.FC = () => {
             initialValues={{
               email: '',
               password: '',
-              confirmPassword: '',
             }}
             validate={validate}
             validateOnChange={false}
@@ -128,31 +102,12 @@ const UserSignUp: React.FC = () => {
             {({ errors, handleChange }) => {
               return (
                 <Form>
-                  {emailExists ? (
-                    <>
-                      <div className={styles.emailExists}>
-                        Email exists. Try to join with a different email.
-                      </div>
-                      &nbsp;
-                    </>
-                  ) : null}
-                  <Grid container spacing={2}>
-                    {constructRow('Email', 'email', handleChange, errors.email)}
-                    {constructRow(
-                      'Password',
-                      'password',
-                      handleChange,
-                      errors.password
-                    )}
-                    {constructRow(
-                      'Confirm Password',
-                      'confirmPassword',
-                      handleChange,
-                      errors.confirmPassword
-                    )}
+                  <Grid container spacing={4}>
+                    {constructRow('email', handleChange, errors.email)}
+                    {constructRow('password', handleChange, errors.password)}
                     <Grid item xs={4}>
-                      <a className={styles.login} href="/users/signin">
-                        Already Registered? Log in
+                      <a className={styles.login} href="/users/signup">
+                        Not Registered? Sign Up
                       </a>
                     </Grid>
                     <Grid item xs={5}>
@@ -164,7 +119,7 @@ const UserSignUp: React.FC = () => {
                         className={styles.submit}
                         type="submit"
                       >
-                        Create
+                        Log In
                       </Button>
                     </Grid>
                   </Grid>
@@ -178,4 +133,4 @@ const UserSignUp: React.FC = () => {
   );
 };
 
-export default UserSignUp;
+export default UserSignIn;
