+++ pages/users/signup.tsx
@@ -1,24 +1,47 @@
-import { signIn } from 'next-auth/client';
+import { useState } from 'react';
 import { Grid, TextField, Button } from '@material-ui/core';
 import { Formik, Form } from 'formik';
-import SigninSchema from 'interfaces/signin';
+import SignupSchema from 'interfaces/signup';
 import Layout from 'components/Layout';
-import styles from './signin.module.css';
+import { signIn } from 'next-auth/client';
+import styles from '../../styles/users/Signup.module.css';
 
 interface FormValues {
   email: string;
   password: string;
+  confirmPassword: string;
 }
 
 interface ErrorValues {
   email?: string;
   password?: string;
+  confirmPassword?: string;
 }
 
-// UserSignIn page. Will need additional email verification to be able to create organizations.
-const UserSignIn: React.FC = () => {
+// UserSignUp page. Will need additional email verification to be able to create organizations.
+const UserSignUp: React.FC = () => {
+  const [emailExists, setEmailExists] = useState(false);
+
   const handleSubmit = async (values: FormValues): Promise<any> => {
-    // Sign in with credentials
+    // Make post request
+    const res: any = await fetch('/api/users/', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        email: values.email,
+        password: values.password,
+      }),
+    });
+
+    // Check if email exists
+    if (!res) {
+      setEmailExists(true);
+      return new Error('Email already exists.');
+    }
+
+    // Sign in user
     try {
       signIn('credentials', {
         email: values.email,
@@ -28,12 +51,13 @@ const UserSignIn: React.FC = () => {
     } catch {
       throw new Error('Could not sign in.');
     }
+    return res;
   };
 
   // Validates inputs
   const validate = (values: FormValues): ErrorValues => {
     const errors: { [k: string]: string } = {};
-    const { error } = SigninSchema.validate(values, {
+    const { error } = SignupSchema.validate(values, {
       context: { strict: true },
       abortEarly: false,
     });
@@ -49,13 +73,14 @@ const UserSignIn: React.FC = () => {
   };
 
   const constructRow = (
+    title: string,
     varName: string,
     handleChange: any,
     error?: string
   ): any => (
     <>
       <Grid item xs={4}>
-        <div className={styles.entryName}>{varName}</div>
+        <div className={styles.entryName}>* {title}</div>
       </Grid>
       <Grid item xs={5}>
         <TextField
@@ -66,7 +91,7 @@ const UserSignIn: React.FC = () => {
           variant="outlined"
           onChange={handleChange}
           id={varName}
-          label={varName}
+          label={title}
           type={
             varName === 'password' || varName === 'confirmPassword'
               ? 'password'
@@ -79,11 +104,11 @@ const UserSignIn: React.FC = () => {
   );
 
   return (
-    <Layout title="Sign In">
+    <Layout title="Sign Up">
       <div className={styles.wrapper}>
         <div className={styles.titles}>
-          <h1>Welcome Back!</h1>
-          <h2>Organization Log In</h2>
+          <h1>Join Us!</h1>
+          <h2>Add your organization!</h2>
           &nbsp;
         </div>
         <div className={styles.entries}>
@@ -91,6 +116,7 @@ const UserSignIn: React.FC = () => {
             initialValues={{
               email: '',
               password: '',
+              confirmPassword: '',
             }}
             validate={validate}
             validateOnChange={false}
@@ -102,12 +128,31 @@ const UserSignIn: React.FC = () => {
             {({ errors, handleChange }) => {
               return (
                 <Form>
-                  <Grid container spacing={4}>
-                    {constructRow('email', handleChange, errors.email)}
-                    {constructRow('password', handleChange, errors.password)}
+                  {emailExists ? (
+                    <>
+                      <div className={styles.emailExists}>
+                        Email exists. Try to join with a different email.
+                      </div>
+                      &nbsp;
+                    </>
+                  ) : null}
+                  <Grid container spacing={2}>
+                    {constructRow('Email', 'email', handleChange, errors.email)}
+                    {constructRow(
+                      'Password',
+                      'password',
+                      handleChange,
+                      errors.password
+                    )}
+                    {constructRow(
+                      'Confirm Password',
+                      'confirmPassword',
+                      handleChange,
+                      errors.confirmPassword
+                    )}
                     <Grid item xs={4}>
-                      <a className={styles.login} href="/users/signup">
-                        Not Registered? Sign Up
+                      <a className={styles.login} href="/users/signin">
+                        Already Registered? Log in
                       </a>
                     </Grid>
                     <Grid item xs={5}>
@@ -119,7 +164,7 @@ const UserSignIn: React.FC = () => {
                         className={styles.submit}
                         type="submit"
                       >
-                        Log In
+                        Create
                       </Button>
                     </Grid>
                   </Grid>
@@ -133,4 +178,4 @@ const UserSignIn: React.FC = () => {
   );
 };
 
-export default UserSignIn;
+export default UserSignUp;
